+++ Mobile/.expo/types/router.d.ts
@@ -6,7 +6,7 @@ export * from 'expo-router';
 declare module 'expo-router' {
   export namespace ExpoRouter {
     export interface __routes<T extends string = string> extends Record<string, unknown> {
-      StaticRoutes: `/` | `/(auth)` | `/(auth)/Login` | `/(auth)/SignUp` | `/(auth)/forgotpassword` | `/(main)` | `/(main)/` | `/(main)/Logout` | `/(main)/Orders` | `/(main)/Orders/Edit` | `/(main)/Orders/Index` | `/(main)/Orders/Invoice` | `/(main)/Products` | `/(main)/Products/` | `/(main)/Products/(Admin)` | `/(main)/Products/(Admin)/` | `/(main)/Products/(Admin)/add` | `/(main)/Products/(Admin)/detail` | `/(main)/Products/(Admin)/update` | `/(main)/Products/(Supplier)` | `/(main)/Products/(Supplier)/Cart` | `/(main)/Products/(Supplier)/invoice` | `/(main)/Products/(Supplier)/orderConfirm` | `/(main)/Products/(Supplier)/productDetail` | `/(main)/Products/(Supplier)/supplierIndex` | `/(main)/Products/Cart` | `/(main)/Products/add` | `/(main)/Products/detail` | `/(main)/Products/invoice` | `/(main)/Products/orderConfirm` | `/(main)/Products/productDetail` | `/(main)/Products/supplierIndex` | `/(main)/Products/update` | `/(main)/Profile` | `/(main)/Profile/` | `/(main)/Profile/Delete` | `/(main)/Profile/Detail` | `/(main)/Profile/Edit` | `/(main)/Settings` | `/(main)/Suppliers` | `/(main)/Suppliers/` | `/(main)/Suppliers/add` | `/(main)/Suppliers/supplierView` | `/(main)/Suppliers/update` | `/(main)/Users` | `/(main)/Users/` | `/(main)/Users/UserProfile` | `/(main)/Users/addUsers` | `/(main)/Users/updateUser` | `/(main)/Workers` | `/(main)/Workers/` | `/(main)/Workers/AttendaceReport` | `/(main)/Workers/EmployeeAttendace` | `/(main)/Workers/EmployeeDetail` | `/(main)/Workers/EmployeesList` | `/(main)/Workers/GenerateReports` | `/(main)/Workers/Summary` | `/(main)/Workers/UpdateEmployee` | `/(main)/Workers/addOvertime` | `/(main)/Workers/addWorkers` | `/(main)/Workers/attendaceModal` | `/(main)/Workers/attendaceScreen` | `/(main)/Workers/employeeOvertime` | `/(main)/Workers/summaryReport` | `/Login` | `/Logout` | `/Orders` | `/Orders/Edit` | `/Orders/Index` | `/Orders/Invoice` | `/Products` | `/Products/` | `/Products/(Admin)` | `/Products/(Admin)/` | `/Products/(Admin)/add` | `/Products/(Admin)/detail` | `/Products/(Admin)/update` | `/Products/(Supplier)` | `/Products/(Supplier)/Cart` | `/Products/(Supplier)/invoice` | `/Products/(Supplier)/orderConfirm` | `/Products/(Supplier)/productDetail` | `/Products/(Supplier)/supplierIndex` | `/Products/Cart` | `/Products/add` | `/Products/detail` | `/Products/invoice` | `/Products/orderConfirm` | `/Products/productDetail` | `/Products/supplierIndex` | `/Products/update` | `/Profile` | `/Profile/` | `/Profile/Delete` | `/Profile/Detail` | `/Profile/Edit` | `/Settings` | `/SignUp` | `/Suppliers` | `/Suppliers/` | `/Suppliers/add` | `/Suppliers/supplierView` | `/Suppliers/update` | `/Users` | `/Users/` | `/Users/UserProfile` | `/Users/addUsers` | `/Users/updateUser` | `/Workers` | `/Workers/` | `/Workers/AttendaceReport` | `/Workers/EmployeeAttendace` | `/Workers/EmployeeDetail` | `/Workers/EmployeesList` | `/Workers/GenerateReports` | `/Workers/Summary` | `/Workers/UpdateEmployee` | `/Workers/addOvertime` | `/Workers/addWorkers` | `/Workers/attendaceModal` | `/Workers/attendaceScreen` | `/Workers/employeeOvertime` | `/Workers/summaryReport` | `/_sitemap` | `/forgotpassword` | `/splash`;
+      StaticRoutes: `/` | `/(auth)` | `/(auth)/Login` | `/(auth)/SignUp` | `/(auth)/forgotpassword` | `/(auth)\Login` | `/(auth)\SignUp` | `/(main)` | `/(main)/` | `/(main)/Logout` | `/(main)/Orders` | `/(main)/Orders/Edit` | `/(main)/Orders/Index` | `/(main)/Orders/Invoice` | `/(main)/Products` | `/(main)/Products/` | `/(main)/Products/(Admin)` | `/(main)/Products/(Admin)/` | `/(main)/Products/(Admin)/add` | `/(main)/Products/(Admin)/detail` | `/(main)/Products/(Admin)/update` | `/(main)/Products/(Supplier)` | `/(main)/Products/(Supplier)/Cart` | `/(main)/Products/(Supplier)/invoice` | `/(main)/Products/(Supplier)/orderConfirm` | `/(main)/Products/(Supplier)/productDetail` | `/(main)/Products/(Supplier)/supplierIndex` | `/(main)/Products/Cart` | `/(main)/Products/add` | `/(main)/Products/detail` | `/(main)/Products/invoice` | `/(main)/Products/orderConfirm` | `/(main)/Products/productDetail` | `/(main)/Products/supplierIndex` | `/(main)/Products/update` | `/(main)/Profile` | `/(main)/Profile/` | `/(main)/Profile/Delete` | `/(main)/Profile/Detail` | `/(main)/Profile/Edit` | `/(main)/Settings` | `/(main)/Suppliers` | `/(main)/Suppliers/` | `/(main)/Suppliers/add` | `/(main)/Suppliers/supplierView` | `/(main)/Suppliers/update` | `/(main)/Users` | `/(main)/Users/` | `/(main)/Users/UserProfile` | `/(main)/Users/addUsers` | `/(main)/Users/updateUser` | `/(main)/Workers` | `/(main)/Workers/` | `/(main)/Workers/AttendaceReport` | `/(main)/Workers/EmployeeAttendace` | `/(main)/Workers/EmployeeDetail` | `/(main)/Workers/EmployeesList` | `/(main)/Workers/GenerateReports` | `/(main)/Workers/Summary` | `/(main)/Workers/UpdateEmployee` | `/(main)/Workers/addOvertime` | `/(main)/Workers/addWorkers` | `/(main)/Workers/attendaceModal` | `/(main)/Workers/attendaceScreen` | `/(main)/Workers/employeeOvertime` | `/(main)/Workers/summaryReport` | `/..\Styles\Auth\forgotPassword` | `/..\Styles\Auth\login` | `/..\Styles\Auth\sighnUp` | `/..\Styles\User\user` | `/Login` | `/Logout` | `/Orders` | `/Orders/Edit` | `/Orders/Index` | `/Orders/Invoice` | `/Products` | `/Products/` | `/Products/(Admin)` | `/Products/(Admin)/` | `/Products/(Admin)/add` | `/Products/(Admin)/detail` | `/Products/(Admin)/update` | `/Products/(Supplier)` | `/Products/(Supplier)/Cart` | `/Products/(Supplier)/invoice` | `/Products/(Supplier)/orderConfirm` | `/Products/(Supplier)/productDetail` | `/Products/(Supplier)/supplierIndex` | `/Products/Cart` | `/Products/add` | `/Products/detail` | `/Products/invoice` | `/Products/orderConfirm` | `/Products/productDetail` | `/Products/supplierIndex` | `/Products/update` | `/Profile` | `/Profile/` | `/Profile/Delete` | `/Profile/Detail` | `/Profile/Edit` | `/Settings` | `/SignUp` | `/Suppliers` | `/Suppliers/` | `/Suppliers/add` | `/Suppliers/supplierView` | `/Suppliers/update` | `/Users` | `/Users/` | `/Users/UserProfile` | `/Users/addUsers` | `/Users/updateUser` | `/Workers` | `/Workers/` | `/Workers/AttendaceReport` | `/Workers/EmployeeAttendace` | `/Workers/EmployeeDetail` | `/Workers/EmployeesList` | `/Workers/GenerateReports` | `/Workers/Summary` | `/Workers/UpdateEmployee` | `/Workers/addOvertime` | `/Workers/addWorkers` | `/Workers/attendaceModal` | `/Workers/attendaceScreen` | `/Workers/employeeOvertime` | `/Workers/summaryReport` | `/_sitemap` | `/forgotpassword` | `/splash`;
       DynamicRoutes: never;
       DynamicRouteTemplate: never;
     }
