+++ Mobile/app/(auth)/Login.tsx
@@ -3,7 +3,7 @@ import { View, Text, TextInput, TouchableOpacity, StyleSheet, Alert } from "reac
 import { useRouter } from "expo-router";
 import { useAuth } from "../../contexts/AuthContext";
 import { Ionicons } from "@expo/vector-icons";
-
+import styles from "@/Styles/Auth/login";
 export default function LoginScreen() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
@@ -99,100 +99,3 @@ export default function LoginScreen() {
         </View>
     );
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        padding: 20,
-    },
-    formContainer: {
-        borderRadius: 20,
-        padding: 20,
-    },
-    logoContainer: {
-        alignItems: 'center',
-        marginBottom: 20,
-    },
-    title: {
-        fontSize: 28,
-        fontWeight: 'bold',
-        color: '#000',
-        textAlign: 'center',
-        marginBottom: 30,
-    },
-    inputContainer: {
-        marginBottom: 20,
-        flexDirection: 'row',
-        alignItems: 'center',
-        backgroundColor: '#fff',
-        borderRadius: 10,
-        borderWidth: 1,
-        borderColor: '#555',
-    },
-    inputIcon: {
-        padding: 10,
-        marginLeft: 5,
-    },
-    input: {
-        flex: 1,
-        padding: 15,
-        fontSize: 16,
-        color: '#000',
-    },
-    eyeIcon: {
-        padding: 10,
-        marginRight: 5,
-        borderLeftColor: "black",
-        borderLeftWidth: 1
-    },
-    loginButton: {
-        backgroundColor: '#000',
-        padding: 15,
-        borderRadius: 10,
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'center',
-        marginTop: 10,
-    },
-    buttonIcon: {
-        marginRight: 10,
-    },
-    loginButtonText: {
-        color: '#fff',
-        fontSize: 18,
-        fontWeight: 'bold',
-    },
-    dividerContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        marginVertical: 20,
-    },
-    divider: {
-        flex: 1,
-        height: 1,
-        backgroundColor: '#555',
-    },
-    dividerText: {
-        marginHorizontal: 10,
-        color: '#000',
-        fontWeight: 'bold',
-        alignSelf:"center"
-    },
-    createAccountText: {
-        color: '#000',
-        fontSize: 14,
-        textAlign: 'center',
-        marginTop: 15,
-       
-    },
-    forgotPasswordContainer: {
-        marginTop: 15,
-    },
-    forgotPasswordText: {
-        color: '#555',
-        fontSize: 14,
-        textAlign: 'center',
-        textDecorationLine: 'underline',
-    },
-});
