+++ Mobile/app/(auth)/_layout.tsx
@@ -1,4 +1,3 @@
-// app/(auth)/_layout.tsx
 import React from "react";
 import { Stack } from "expo-router";
 import { useAuth } from "@/contexts/AuthContext";
