+++ Mobile/tsconfig.json
@@ -15,6 +15,7 @@
   "include": [
     "**/*.ts",
     "**/*.tsx",
+    "**/*.js",
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
   ]
