--- o.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react'
-import { Text, View } from 'react-native'
-
-export class o extends Component {
-  render() {
-    return (
-      <View>
-        <Text> textInComponent </Text>
-      </View>
-    )
-  }
-}
-
-export default o
